requirements.txt
dj-database-url==0.4.2
Django==1.11.2
django-webpack-loader==0.5.0
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
whitenoise==3.3.0
